\name{CRU.STATIONS.URL}
\alias{CRU.STATIONS.URL}
\docType{data}
\title{A url for the CRU station data
 
}
\description{url of the CRU station data
 
}
\usage{CRU.STATIONS.URL}
\format{
  The format is:
 chr "http://www.cru.uea.ac.uk/cru/data/temperature/station-data/station-data.zip"
}
\details{Provides a link to the CRU station data for download
 
}
\source{\url{http://www.cru.uea.ac.uk/cru/data/temperature/station-data}
 
}
\references{Brohan, P., Kennedy, J., Harris, I., Tett, S.F.B. and Jones, P.D., 2006: 
      Uncertainty estimates in regional and global observed temperature changes:
      a new dataset from 1850. J. Geophys. Res. 111, D12106
 
 
}
\examples{
print(CRU.STATIONS.URL)
 
}
\keyword{datasets}
