\name{CRUTEMP3.RESULTS.URL}
\alias{CRUTEMP3.RESULTS.URL}
\docType{data}
\title{A url to Crutemp3 temperature series
 
}
\description{This is the url to the crutemp3 land temperature record.
  The data consists of monthly anomalies relative to a 1961-1990
  baseline. The data is starts in 1850 and is updated monthly.
  
 
}
\usage{ CRUTEMP3.RESULTS.URL }
\format{
  The format is:
 chr "http://www.cru.uea.ac.uk/cru/data/temperature/crutem3gl.txt"
}
\details{The dataset contains monthly figures for anomalies and annual
  figures as well. In addition, a metric for the percentage of land
  sampled is included
 
}
\source{the data is described here
  \url{http://www.cru.uea.ac.uk/cru/data/temperature/}
 
}
\references{documentation for the data is contained the following paper
  \url{http://www.cru.uea.ac.uk/cru/data/temperature/HadCRUT3_accepted.pdf}
 
}
\examples{
 
print(CRUTEMP3.RESULTS.URL)  
}
\keyword{datasets}
