\name{HADSST2.RESULTS.URL}
\alias{HADSST2.RESULTS.URL}
\docType{data}
\title{A url to Hadley SST temperature series
 
}
\description{This is the url to the Hadley SST temperature record.
  The data consists of monthly anomalies relative to a 1961-1990
  baseline. The data is starts in 1850 and is updated monthly.
 
}
\usage{ HADSST2.RESULTS.URL }
\format{
  The format is:
 chr "http://www.cru.uea.ac.uk/cru/data/temperature/hadsst2gl.txt"
}
\details{The dataset contains monthly figures for anomalies and annual
  figures as well.  
 
}
\source{the data is described here
  \url{http://www.cru.uea.ac.uk/cru/data/temperature/}
 
}
 
\examples{
print(HADSST2.RESULTS.URL)
 
}
\keyword{datasets}
