\name{isMts}
\alias{isMts}
 
\title{A function to test for class type
 
}
\description{Checks whether an object is an mts or multiple time
  series class. 
 
}
\usage{isMts(data)
}
 
\arguments{
  \item{data}{a data object to be tested
 
}
}
\details{This function simply makes code more readable as opposed
  to \code{class(data)[1]}
 
}
\value{return true or false and throws a warning if false
 
}
 
\author{Steven Mosher
 
}
 

 
 
\keyword{ Internals }
 
