\name{V3.MEAN.ADJ.URL}
\alias{V3.MEAN.ADJ.URL}
\docType{data}
\title{ Url for Adjusted GHCN V3 data
 
}
\description{ The url to the Adjusted GHCN V3 data.
  This dataset is an Adjusted dataset of Tave or
  Average temperature.
 
}
\usage{V3.MEAN.ADJ.URL}
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/ghcnm.tavg.latest.qca.tar.gz"
}
 
\source{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/}
 
}
\references{ \url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/README}
 
}
\examples{
data(V3.MEAN.ADJ.URL)
print(V3.MEAN.ADJ.URL)  
}
\keyword{datasets}
