\name{V3.MEAN.RAW.URL}
\alias{V3.MEAN.RAW.URL}
\docType{data}
\title{Url for UnAdjusted GHCN V3 data
 
}
\description{The url to the UnAdjusted GHCN V3 data.
  This dataset is an Adjusted dataset of Tave or
  Average temperature.
 
}
\usage{V3.MEAN.RAW.URL}
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/ghcnm.tavg.latest.qcu.tar.gz"
}
 
\source{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/}
 
}
\references{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/README}
 
}
\examples{
data(V3.MEAN.RAW.URL)
print(V3.MEAN.RAW.URL)  
}
\keyword{datasets}
