\name{V3.TMAX.ADJ.URL}
\alias{V3.TMAX.ADJ.URL}
\docType{data}
\title{Url for  Adjusted GHCN V3 TMAX data
 
}
\description{A url to the GHCN V3 TMAX dataset. This data
  is adjusted. see the raw version of the URL for unadjusted data
 
}
\usage{V3.TMAX.ADJ.URL}
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/ghcnm.tmax.latest.qca.tar.gz"
}
 
\source{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/}
 
}
\references{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/README}
 
}
\examples{
data(V3.TMAX.ADJ.URL)
print(V3.TMAX.ADJ.URL)  
}
\keyword{datasets}
