\name{V3.TMAX.RAW.URL}
\alias{V3.TMAX.RAW.URL}
\docType{data}
\title{ A url for the GHCN V3 Tmax data, unadjusted
 
}
\description{This url points to the unadjusted Tmax (maximum)
  dataset for GHCN V3
 
}
\usage{V3.TMAX.RAW.URL}
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/ghcnm.tmax.latest.qcu.tar.gz"
}
 
\source{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/}
 
}
\references{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/README}
 
}
\examples{
data(V3.TMAX.RAW.URL)
print(V3.TMAX.RAW.URL)  
}
\keyword{datasets}
