\name{V3.TMIN.RAW.URL}
\alias{V3.TMIN.RAW.URL}
\docType{data}
\title{A url to the GHCN V3 undjusted Tmin data
 
}
\description{A url to the unadjusted Tmin (minimum)
  Temperature dataset for GHCN version 3.
  
 
}
\usage{V3.TMIN.RAW.URL}
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/ghcnm.tmin.latest.qcu.tar.gz"
}
 
\source{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/}
 
}
\references{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/v3/README}
 
}
\examples{
data(V3.TMIN.RAW.URL)
print(V3.TMIN.RAW.URL)  
}
\keyword{datasets}
